generate correct schema if parent class has only a single child
Parent classes that only have a single child (e.g. because they are open for extension, but not yet have more childs) are generating a wrong schema. The emitted oneOf element has only a single $ref, which is not valid. It needs at least two array elements. This is changed, so that oneOf is only emitted if more than one child is available.